+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Features:
   + Added FDEPL flag "Internal" for marking classes, methods and fields as package-private in Java
+  + Added FDEPL flag "Internal" for marking classes, methods and fields as internal in Swift
 - Bug fixes:
   + Fixed compilation issues for Maps with Enums as keys in iOS generator when a non-default value
     is specified for "C++ internal namespace" command line parameter.